+++ com.unity.ml-agents/CHANGELOG.md
@@ -9,6 +9,9 @@ and this project adheres to
 ### Minor Changes
 #### com.unity.ml-agents (C#)
 - Upgrade to 2.0.1
+#### ml-agents / ml-agents-envs / gym-unity (Python)
+- Set gym version in gym-unity to gym release 0.20.0
+- Added minimal analytics collection to LL-API (#5511)
 
 ## [2.0.0] - 2021-09-01
 ### Minor Changes